create new separate class for FTPopOverMenuCell
